added controller for tests
